requirements.txt
Pillow
pytesseract
XlsxWriter